Start ripping out components, delay adding networked scene until entry
