+++ src/room.html
@@ -18,11 +18,6 @@
 
     <a-scene
         physics
-        networked-scene="adapter: janus;
-                         audio: true;
-                         debug: true;
-                         onConnect: App.onConnect;
-                         connectOnLoad: false;"
         mute-mic="eventSrc: a-scene; toggleEvents: action_mute"
         2d-mute-state-indicator
         light="defaultLightsEnabled: false">
